Move the code to utils, #1438
